Crate cranelift_simplejit

Top-level lib.rs for cranelift_simplejit.

Structs

SimpleJITBackend
A SimpleJITBackend implements Backend and emits code and data into memory where it can be directly called and accessed.

SimpleJITBuilder
A builder for SimpleJITBackend.

Constants

VERSION
Version number of this crate.