Struct cranelift_simplejit::SimpleJITBackend

pub struct SimpleJITBackend { /* fields omitted */ }

A SimpleJITBackend implements Backend and emits code and data into memory where it can be directly called and accessed.

See the SimpleJITBuilder for a convenient way to construct SimpleJITBackend instances.