Struct cranelift_simplejit::SimpleJITModule

pub struct SimpleJITModule { /* fields omitted */ }

A SimpleJITModule implements Module and emits code and data into memory where it can be directly called and accessed.

See the SimpleJITBuilder for a convenient way to construct SimpleJITModule instances.

Implementations

impl SimpleJITModule

pub unsafe fn free_memory(&mut self)

Free memory allocated for code and data segments of compiled functions.

Safety

Because this function invalidates any pointers retrived from the corresponding module, it should only be used when none of the functions from that module are currently executing and none of the fn pointers are called afterwards.

pub fn get_finalized_function(&self, func_id: FuncId) -> *const u8

Returns the address of a finalized function.

pub fn get_finalized_data(&self, data_id: DataId) -> (*const u8, usize)

Returns the address and size of a finalized data object.

pub fn finalize_definitions(&mut self)

Finalize all functions and data objects that are defined but not yet finalized. All symbols referenced in their bodies that are declared as needing a definition must be defined by this point.

Use get_finalized_function and get_finalized_data to obtain the final artifacts.

pub fn new(builder: SimpleJITBuilder) -> Self

Create a new SimpleJITModule.